// src/app/Model/employee.ts
export class Employee {


        id:  any;
	firstName: any;
	lastName: any;
	cin: any;
	age: any;   //kenet number

	
	/*
	positions: string;
        invitation: string;
        photo: string;
	
        username: string;
        email: string;
        password: string;
        phoneNumber: string;
        role: string;
*/

}